Fixes after merge (#79)
* Fixed RideCreateDTO Bug

* Nothing

* Final changes

* Final Changes

* Search autocomplete + things deleted by merge added + style fixes.

* Background color changed.

+++ blablacar-service/src/main/java/cz/fi/muni/pa165/teamred/service/facade/RideFacadeImpl.java
@@ -70,8 +70,12 @@ public class RideFacadeImpl implements RideFacade {
 
     @Override
     public void deleteRide(Long rideId) {
-        Ride ride = new Ride();
-        ride.setId(rideId);
+        Ride ride = rideService.findById(rideId);
+
+        if(ride == null) {
+            throw new IllegalArgumentException();
+        }
+
         rideService.deleteRide(ride);
     }
